Add COPY_CHECK environment variable
The COPY_CHECK environment variable provides a check to see if we're
allowed to execute the copy command.  The COPY_TASK_UUID is replaced
with the sparql_escaped UUID of the copy task.  The URIs of the
attributes used here should be changed into something sensible.  This
code will be altered to support the fetching/writing from/to multiple
stores.